# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Fuse — Federated Query Engine API
  description: REST API for the Fuse cross-datasource federated query engine.
  version: 2.0.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://fuse.huanji.profile.aws.dev
    description: Playground (Amazon VPN only)
  - url: http://localhost:9400
    description: Local development
paths:
...
  /api/fuse/query:
    post:
      summary: Execute a federated query
      description: |
        Parses a SQL or PPL query, routes sub-queries to the appropriate
        connectors, and returns merged results.

        The query must reference tables as `datasource.table` (e.g., `cluster_a.logs`).
      operationId: query
      tags: [Query]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
            examples:
              sql:
                summary: SQL query
                value:
                  query: "SELECT * FROM cluster_a.logs WHERE status = 500"
                  format: sql
              ppl:
                summary: PPL query
                value:
                  query: "source = cluster_a.logs | where status >= 500 | stats count() by service"
                  format: ppl
      responses:
        "200":
          description: Query results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  schemas:
    QueryRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          description: SQL or PPL query string
        format:
          type: string
          enum: [sql, ppl]
          default: sql
          description: Query language format
        analyze:
          type: boolean
          default: false
          description: When true, includes execution_profile with actual timing in response
        timeout_ms:
          type: integer
          default: 30000
          description: Per-query timeout in milliseconds
        result_format:
          type: string
          enum: [json, csv]
          default: json
          description: Response format
        params:
          type: object
          additionalProperties: {}
          description: Named parameters for $name placeholders
        page_size:
          type: integer
          description: Rows per page (enables cursor pagination). When set, response includes next_cursor if more rows exist.
        cursor:
          type: string
          description: Cursor from previous response's next_cursor field for fetching the next page.
    QueryResponse:
      type: object
      properties:
        columns:
          type: array
          items:
            type: string
          description: Column names (includes _datasource for cross-datasource queries)
        rows:
          type: array
          items:
            type: array
            items: {}
          description: Row data (array of arrays)
        metadata:
          type: object
          properties:
            total_rows:
              type: integer
            format:
              type: string
            datasources_queried:
              type: array
              items:
                type: string
              nullable: true
              description: List of datasource IDs involved in the query
            datasource_stats:
              type: object
              additionalProperties:
                $ref: "#/components/schemas/DatasourceStat"
              nullable: true
              description: Per-datasource row counts and latency
            execution_profile:
              $ref: "#/components/schemas/ExecutionProfile"
              nullable: true
              description: Present when analyze=true
        next_cursor:
          type: string
          nullable: true
          description: Cursor for fetching the next page. Present when page_size is set and more rows exist.
        partial_errors:
          type: array
          items:
            $ref: "#/components/schemas/PartialError"
          description: Errors from individual datasources in UNION ALL (other sources still return data)
    DatasourceStat:
      type: object
      properties:
        rows:
          type: integer
        latency_ms:
          type: integer
    ExecutionProfile:
      type: object
      properties:
        total_ms:
          type: integer
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/ProfileNode"
    ProfileNode:
      type: object
      properties:
        op:
          type: string
          description: Operation name (RemoteScan, Merge, HashJoin, etc.)
        datasource:
          type: string
          nullable: true
        actual_rows:
          type: integer
        actual_ms:
          type: integer
        data_bytes:
          type: integer
          nullable: true
        pushdown:
          type: array
          items:
            type: string
          description: Pushdown operations (filter, projection, limit)
        children:
          type: array
          items:
            $ref: "#/components/schemas/ProfileNode"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    PartialError:
      type: object
      properties:
        datasource:
          type: string
        error:
          type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...